+++ skbio/io/format/embed.py
